Cart Operations

+++ E_Commerse/src/main/java/org/project/models/Products.java
@@ -3,12 +3,12 @@ package org.project.models;
 public class Products {
     private int id;
     private String name;
-    private double price;
+    private int price;
     private int quantity;
     private int categoryId;
     public Products() {}
 
-    public Products(int id, String name, double price, int quantity, int categoryId) {
+    public Products(int id, String name, int price, int quantity, int categoryId) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -32,11 +32,11 @@ public class Products {
         this.name = name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
